Move Rustdoc-related code into its own module
