+++ crates/html_to_markdown/src/structure.rs
@@ -1,0 +1,1 @@
+pub mod rustdoc;
